+++ src/indexes.ts
@@ -1,29 +1,49 @@
 export const JS_INDEX_URL =
     "https://github.com/vemel/github_actions_js/tree/main/nodejs_workflows";
-export const PY_INDEX_URL =
-    "https://github.com/vemel/github_actions_js/tree/main/python_workflows";
 
 export interface IIndex {
     url: string;
-    shortcut?: string;
+    title: string;
+    shortcut: string;
 }
 
-export const shortcuts = {
-    node: JS_INDEX_URL,
-    python: PY_INDEX_URL
-};
+export const INDEXES: Array<IIndex> = [
+    {
+        url: JS_INDEX_URL,
+        title: "Node.js",
+        shortcut: "node"
+    },
+    {
+        url:
+            "https://github.com/vemel/github_actions_js/tree/main/python_workflows",
+        title: "Python",
+        shortcut: "python"
+    },
+    {
+        url:
+            "https://github.com/actions-rs/example/tree/master/.github/workflows",
+        title: "Rust",
+        shortcut: "rust"
+    }
+];
 
-export const INDEXES: Array<string> = [JS_INDEX_URL, PY_INDEX_URL];
-
-export function getIndexResource(url: string): string {
-    if (shortcuts[url]) return shortcuts[url];
-    return url;
-}
-
-export function getShortcut(url: string): string {
-    for (const shortcut of Object.keys(shortcuts)) {
-        const shortcutURL = shortcuts[shortcut];
-        if (shortcutURL === url) return shortcut;
+export function getIndexURL(url: string): string {
+    for (const index of INDEXES) {
+        if (index.url === url || index.shortcut === url) return index.url;
     }
     return url;
+}
+
+export function getIndexResource(url: string): IIndex | null {
+    for (const index of INDEXES) {
+        if (index.url === url || index.shortcut === url) return index;
+    }
+    return null;
+}
+
+export function getShortcut(url: string): string | null {
+    for (const index of INDEXES) {
+        if (index.url === url || index.shortcut === url) return index.shortcut;
+    }
+    return null;
 }