+++ src/runCheck.ts
@@ -55,7 +55,9 @@ export async function runCheck(
 
     const newWorkflow = new Merger(true).merge(localWorkflow, remoteWorkflow);
     if (removeMarker) {
-        newWorkflow.job.steps.forEach(step => step.makeNonManaged());
+        newWorkflow.jobs.forEach(job =>
+            job.steps.forEach(step => step.makeNonManaged())
+        );
     }
     return checker.getChecks(newWorkflow);
 }