+++ src/workflow/merger.ts
@@ -13,27 +13,35 @@ export class Merger {
         if (this.force) current.triggers = update.triggers;
     }
 
-    mergeJob(current: Workflow, update: Workflow): void {
-        const currentJob = current.job;
-        const updateJob = update.job;
+    mergeJobs(current: Workflow, update: Workflow): void {
+        update.jobNames.forEach(jobName => {
+            const updateJob = update.getJob(jobName);
+            if (!current.jobNames.includes(jobName)) {
+                current.setJob(updateJob.clone());
+                return;
+            }
 
-        if (this.force) currentJob.env = updateJob.env;
-        if (this.force) currentJob.runsOn = updateJob.runsOn;
-        if (this.force) currentJob.runsIf = updateJob.runsIf;
-        if (this.force) currentJob.strategy = updateJob.strategy;
-    }
+            const currentJob = current.getJob(jobName);
+            if (this.force) currentJob.env = updateJob.env;
+            if (this.force) currentJob.runsOn = updateJob.runsOn;
+            if (this.force) currentJob.runsIf = updateJob.runsIf;
+            if (this.force) currentJob.strategy = updateJob.strategy;
 
-    mergeSteps(current: Workflow, update: Workflow): void {
-        const currentJob = current.job;
-        const updateSteps = update.job.steps.map(step => step.makeManaged());
-        currentJob.steps = currentJob.mergeSteps(updateSteps);
+            const updateSteps = updateJob.steps.map(step => step.makeManaged());
+            currentJob.steps = currentJob.mergeSteps(updateSteps);
+        });
+        if (this.force)
+            current.jobNames
+                .filter(jobName => !update.jobNames.includes(jobName))
+                .map(jobName => {
+                    current.deleteJob(jobName);
+                });
     }
 
     merge(current: Workflow, update: Workflow): Workflow {
         const result = current.clone();
         this.mergeWorkflow(result, update);
-        this.mergeJob(result, update);
-        this.mergeSteps(result, update);
+        this.mergeJobs(result, update);
         return result;
     }
 }